Bug 1981465: Assisted installer wait for ready master nodes on bootstrap kube-apiserver though the kube-apiserver moved to one of the masters On the bootstrap node the assisted-installer use the loopback-kubeconfig to query the kube-apiserver for the number of ready master nodes.
Usually both master nodes join the cluster and become ready before
bootkube takes down the bootstrap control plane so the loopback kubeconfig works.
But in case clusteer bootstrap finish before the 2 master nodes
are ready the assisted-installer will wait forever since the it's
using the loopback-kubeconfig and the bootstrap control plane is down
resulting in "connection refused"

The assisted-installer should query the kube-apiserver running on
one of the master nodes, for that to work it should use the real
kubeconfig instead of the loopback kubeconfig.